cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-firestore==2.*
pytz==2023.*